Refactor avatar to be static image with background instead of div background
