+++ shmoney-react/src/styles/NavBar.css
@@ -25,7 +25,6 @@
   width: 40px;
   height: 40px;
   border-radius: 50%;
-  background-color: #E3B505;
   margin: 0 5px;
 }
 .avatar:hover {
